Development

+++ resources/views/pages/contacts/⚡show/show.php
@@ -9,6 +9,7 @@ new class extends Component
 
     public function mount(User $contact)         //avant de render ( 1x seulement)
     {
+        $this->authorize('view', $contact);        //sinon ça doit à chaque sort vérifier authorization        //tous les users peuvent voir tous les contacts
         $this->contact_id = $contact->id;
     }
 
@@ -17,12 +18,13 @@ new class extends Component
         $contact = \App\Models\User::findOrFail($this->contact_id);
         $user = auth()->user();
 
-        return view('pages.contacts.⚡show.show', ['contact' => $contact, 'user' => $user]);
+        return view('pages.contacts.⚡show.show', ['contact' => $contact, 'user' => $user])->title(__('general.contact_detail'));
     }
 
     public function destroy()
     {
         $contact = User::findOrFail($this->contact_id);
+        $this->authorize('delete', $contact);
         $contact->delete();
         return redirect(route('pages::contacts.index', ['locale' => app()->getLocale()]));
     }
